refactor: extract Material, Ray, tracer, and renderer into dedicated modules
