+++ wasm-lib/src/lib.rs
@@ -1,8 +1,12 @@
 mod entity;
 mod intersection;
+mod material;
 mod post_processing;
+mod ray;
+mod renderer;
 mod rgb;
 mod scene;
+mod tracer;
 mod vec2;
 mod vec3;
 
@@ -10,7 +14,6 @@ use wasm_bindgen::prelude::*;
 
 #[wasm_bindgen]
 extern "C" {
-    pub fn alert(s: &str);
     #[wasm_bindgen(js_namespace = console)]
     pub fn log(s: &str);
 }
